T9814: ensure matomo script isn't added more than once (#91)
Since 1.38 it has been possible that the `SkinAfterBottomScripts` hook could be ran twice. Once is in `OutputPage::tailElement()`, once in `OutputPage::getBottomScripts()`. Since `OutputPage::tailElement()` also runs `OutputPage::getBottomScripts()`, if `OutputPage::tailElement()` is ran, the hook is ran twice and the script here gets added twice. This should prevent that.